Move JSON converter for TensorSplitsCollection
